move files to clarify the code structure
